// src/app/models/ireportformdata.ts
export interface IReportFormData {
  name: string;
  date: string;
  hour: string;
  hr: string;
  rr: string;
  temp: string;
  bpSystolic: string;
  bpDiastolic: string;
  o2: string;
  glucose: string;
  bowel: boolean;
  urine: boolean;
  report: string;
  indications: string;
  signature: string;
}